+++ src/GameObject.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
  */
 public abstract class GameObject { //khai bao abstract de: chi cac class thuoc ve game ke thua va su dung
     //khai bao cac thuoc tinh toa do, Anh nen cua doi tuong trong Game
-    protected int postionX;
-    protected int postionY;
+    protected double postionX;
+    protected double postionY;
     protected BufferedImage sprite;
 }
